% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mult}
\alias{mult}
\title{Multiply two numbers.}
\usage{
mult(n1, n2)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The product of \code{x} and \code{y}.
}
\description{
Multiply two numbers.
}
\examples{
mult(1, 1)
mult(10, 1)
}

